// src/create-user/create-user.component.ts
import { Component, EventEmitter, Output, inject } from '@angular/core';
import { IUser } from '../interfaces/IUser';
import { ReactiveFormsModule, FormBuilder, Validators, FormGroup } from '@angular/forms';
import { HoverDirective } from '../directives/hover.directive';
import { HoverBorderDirective } from '../directives/hover-border.directive';
import { IGradientConfiguration } from '../interfaces/IGradientConfiguration';

@Component({
  selector: 'app-create-user',
  imports: [ReactiveFormsModule, HoverDirective, HoverBorderDirective],
  templateUrl: './create-user.component.html',
  styleUrl: './create-user.component.scss',
})
export class CreateUserComponent {

  @Output() userCreated: EventEmitter<IUser> = new EventEmitter<IUser>();

  private fb: FormBuilder = inject(FormBuilder);

  readonly hoverConfig: IGradientConfiguration = {
    delay: 500,
    colors: ['purple', 'orange', 'cyan', 'red', 'cyan', 'orange', 'purple'],
    thickness: 3
  };

  form: FormGroup = this.fb.nonNullable.group({
    id: [{ value: null, disabled: true }],
    name: ['', [Validators.required, Validators.minLength(2), Validators.maxLength(100)]],
    username: ['', [Validators.required, Validators.minLength(3), Validators.maxLength(30)]],
    email: ['', [Validators.required, Validators.email, Validators.maxLength(100)]],
    phone: ['', [Validators.required, Validators.minLength(10), Validators.maxLength(25)]],
    website: ['', [Validators.maxLength(100)]],
    address: this.fb.nonNullable.group({
      city: ['', [Validators.required, Validators.maxLength(50)]],
      street:['', [Validators.required, Validators.maxLength(100)]],
      suite: ['', [Validators.maxLength(50)]],
      zipcode: ['',[Validators.required, Validators.minLength(5), Validators.maxLength(10)]],
      geo: this.fb.nonNullable.group({
        lat: ['', [Validators.required]],
        lng: ['', [Validators.required]]
      }),
    }),
    company: this.fb.nonNullable.group({
      name: ['', [Validators.required, Validators.maxLength(50)]],
      catchPhrase: ['', [Validators.maxLength(200)]],
      bs: ['', [Validators.maxLength(100)]]
    })
  });

  onSubmit(): void {
    if (this.form.invalid) {
      this.form.markAllAsTouched();
      return;
    }

    const user: IUser = {
      ...this.form.getRawValue(),
      id: Date.now()
    };

    this.userCreated.emit(user);
    this.form.reset();
  }

}
